Trim out MiniPaletteEntry struct and embed entity resource index into colour
This is good because it halves (or better!) the size of the palette data struct, which is a boon for the CPU cache.
This will greatly improve performance.
There is no error with using the ARGB color value both as color and as an embedded value, as an ARGB of 0, 255, x, x is not a sane colour value anyways and wouldn't get used.